[ozone/external_wayland] Refactor for creation and initialization of the gpu window
app_shell doesn't use and doesn't create views widget, but creates window directly from aura as WindowTreeHost.
In this case only ozone platform window creation is called, but initialization as for views widget is not called.
ozone-wayland creates gpu window during initialization of the platform ozone window.
This refactoring moves gpu window creation from initialization to creation of the platform window.

Refactor for creation and initialization of the gpu window:
 1. Creation of the gpu window with creation of the platform ozone window
 2. Initialization of the gpu window with initialization of the platform ozone window