[ozone/external_wayland] Fix events issue for app_shell
App shell is a application which don't call initialize of the platform
wayland window. Initialize of the platform window in ozone-wayland adds
window to a list of the open windows. Therefore in case app shell root window
wasn't added to the list. Events handles for focused window only but the focus can
be set for open window only. Therefore we should add any platform window to list
of the open windows when the window create.